requirements.txt
absl-py==2.1.0
attrs==23.2.0
bcrypt==4.1.2
cffi==1.16.0
contourpy==1.2.0
cycler==0.12.1
flatbuffers==24.3.25
fonttools==4.50.0
importlib_resources==6.4.0
kiwisolver==1.4.5
matplotlib==3.8.3
mediapipe==0.10.9
mysql-connector-python==8.3.0
numpy==1.26.4
opencv-contrib-python==4.9.0.80
packaging==23.2
pillow==10.2.0
protobuf==3.20.3
pycparser==2.21
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sounddevice==0.4.6
zipp==3.18.1
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.36
langchain-experimental==0.0.55
langchain-openai==0.1.1
